+++ articleapp/urls.py
@@ -1,10 +1,8 @@
 from django.urls import path
 
 from articleapp.views import ArticleListView, ArticleDetailView, \
-    ArticleCreateView, ArticleUpdateView, ArticleDeleteView, AddLike, Dislike, search
+    ArticleCreateView, ArticleUpdateView, ArticleDeleteView, AddLike, Dislike
 from mainapp.views import main
-
-
 
 app_name = 'articleapp'
 
@@ -18,5 +16,4 @@ urlpatterns = [
     path('article/<uuid:pk>/edit/', ArticleUpdateView.as_view(), name='edit'),
     path('article/<uuid:pk>/delete/', ArticleDeleteView.as_view(),
          name='delete'),
-    path('search', search, name='search'),
 ]
