+++ articleapp/views.py
@@ -11,20 +11,6 @@ from django.views import View
 from articleapp.forms import ArticleForm
 from articleapp.models import Article
 from mainapp.views import main
-from django.db.models import Q
-
-
-def search(request):
-    if request.method == "POST":
-        searched = request.POST['searched']
-        results = Article.objects.filter(
-            Q(content__contains=searched) | Q(title__contains=searched))
-        return render(request, 'search.html',
-                      {'searched': searched,
-                       'results': results})
-    else:
-        return render(request, 'search.html',
-                      {})
 
 
 # Отображение содержимого из модели Article
