+++ mainapp/templates/mainapp/includes/inc_menu.html
@@ -7,7 +7,7 @@
         </div>
         <form class="form-inline my-2 my-lg-0 d-flex align-items-center"
               method="POST"
-              action="{% url 'article:search' %}">
+              action="{% url 'search:search' %}">
             {% csrf_token %}
           <input class="form-control mr-sm-2"
                  type="search"
