+++ searchapp/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+from searchapp.views import search
+
+
+app_name = 'searchapp'
+
+urlpatterns = [
+    path('search', search, name='search'),
+]
