+++ static/assets/css/style.css
@@ -11,6 +11,10 @@
 body {
   font-family: "Open Sans", sans-serif;
   color: #444444;
+  height:100vh;
+  margin:0;
+  display:flex;
+  flex-direction:column;
 }
 
 a {
@@ -1781,6 +1785,8 @@ section {
   padding: 0 0 30px 0;
   color: #fff;
   font-size: 14px;
+  min-height:85px;
+  margin-top:auto;
 }
 #footer .footer-newsletter {
   padding: 50px 0;
